feat: refactor of UI to use mantine (#290)
A huge refactor. This replaces all uses of EUI in favor of Mantine. This was to remove some of the strange setup required by EUI (themes, icons, etc.)

Additionally contains the work of adding the blocknote (tiptap extended) description parser. Somewhat in a messy state, but works well enough to move forward with at the moment.